+++ jingeong/week_3/gradePoint/Evaluation/Grade.java
@@ -1,3 +1,5 @@
+package Evaluation;
+
 public abstract class Grade {
     abstract String makeGrade();
