+++ jingeong/week_3/gradePoint/Evaluation/Normal.java
@@ -1,17 +1,18 @@
+package Evaluation;
+
 import java.util.HashMap;
 import java.util.LinkedHashMap;
-import java.util.Set;
 
-public class Required extends Grade{
+public class Normal extends Grade{
+
     private final static HashMap<Integer, String> grade = new LinkedHashMap<>();
     private int point;
 
-    public Required(int point) {
-        grade.put(95, "S");
+    public Normal(int point) {
         grade.put(90, "A");
         grade.put(80, "B");
         grade.put(70, "C");
-        grade.put(60, "D");
+        grade.put(55, "D");
         grade.put(0, "F");
 
         this.point = point;
