+++ jingeong/week_3/gradePoint/Evaluation/Required.java
@@ -1,16 +1,18 @@
+package Evaluation;
+
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 
-public class Normal extends Grade{
-
+public class Required extends Grade{
     private final static HashMap<Integer, String> grade = new LinkedHashMap<>();
     private int point;
 
-    public Normal(int point) {
+    public Required(int point) {
+        grade.put(95, "S");
         grade.put(90, "A");
         grade.put(80, "B");
         grade.put(70, "C");
-        grade.put(55, "D");
+        grade.put(60, "D");
         grade.put(0, "F");
 
         this.point = point;
